Oscillator rename
