+++ src/lib.rs
@@ -1,7 +1,7 @@
 extern crate portaudio;
 pub mod analyze;
 pub mod event;
-pub mod new_oscillator;
+pub mod oscillator;
 pub mod portaudio_duplex_setup;
 pub mod portaudio_output_setup;
 pub mod ratios;
